冷提交代码

+++ admin/src/main/java/com/admin/modules/sys/service/impl/CompanyServiceImpl.java
@@ -3,6 +3,7 @@ package com.admin.modules.sys.service.impl;
 import com.admin.common.utils.PageUtils;
 import com.admin.common.utils.Query;
 import com.admin.common.utils.Tools;
+import com.admin.modules.sys.dao.AreaDao;
 import com.admin.modules.sys.dao.CompanyDao;
 import com.admin.modules.sys.entity.CompanyEntity;
 import com.admin.modules.sys.entity.vo.CompanyEntityVo;
@@ -22,6 +23,8 @@ public class CompanyServiceImpl extends ServiceImpl<CompanyDao, CompanyEntity> i
 
     @Autowired
     private CompanyDao dao;
+    @Autowired
+    private AreaDao areaDao;
 
     @Override
     public PageUtils queryPage(Map<String, Object> params) {
@@ -69,6 +72,11 @@ public class CompanyServiceImpl extends ServiceImpl<CompanyDao, CompanyEntity> i
 
     @Override
     public int deleteComById(Integer id) {
+        //删除则校验该公司下面有无绑定的待生效、生效中的合同，如果有则提示：删除失败，该公司有待生效/生效中的合同。
+        int pactcount = areaDao.getPactByCompanyId(id,"tb_company");
+        if (pactcount>0){
+            throw new RuntimeException("删除失败，该公司有待生效/生效中的合同");
+        }
         int count = dao.deleteComById(id);
         return count;
     }
